Create 8_Lambda_Expressions.java
Write the following methods that return a lambda expression performing a specified action:
(i) PerformOperation isOdd(): The lambda expression must return true if a number is odd or false if it is even.
(ii) PerformOperation isPrime(): The lambda expression must return true if a numberis prime or false if it is composite.
(iii) PerformOperation isPalindrome(): The lambda expression must return true if a number is a palindrome or false if it is not.
Write a JAVA program using above lambda expressions to take 2 integers as input where the first integer specifies the condition to check for (case 1 for Odd/Even, case 2 for Prime/Composite, or case 3 for Palindrome). The second integer denotes the number to be checked.